+++ app/server/modules/i18n.js
@@ -36,7 +36,7 @@ const getLangDir = function (_) {
   // Todo: Add `resolvedDirection` to `intl-dom` so can grab that instead here
   const lang = _.resolvedLocale;
   // @ts-expect-error Polyfill for a new standard
-  const {direction} = new Intl.Locale(lang).textInfo;
+  const {direction} = new Intl.Locale(lang).getTextInfo();
   // Don't bother to make default of "ltr" explicit
   const dir = direction === 'rtl' ? direction : undefined;
   return {